@@ -1,7 +1,13 @@
 module websocket-autobahn
 
-go 1.13
+go 1.22.0
 
 replace github.com/pkgz/websocket => ../
 
-require github.com/pkgz/websocket v0.0.0-00010101000000-000000000000
+require (
+	github.com/pkgz/websocket v1.3.0
+	github.com/gobwas/httphead v0.1.0 // indirect
+	github.com/gobwas/pool v0.2.1 // indirect
+	github.com/gobwas/ws v1.4.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+)
